third_party/WebKit/LayoutTests/http/tests/websocket/httponly-cookie-expected.txt

CONSOLE WARNING: Synchronous XMLHttpRequest on the main thread is deprecated because of its detrimental effects to the end user's experience. For more help, check http://xhr.spec.whatwg.org/.
Test WebSocket sends HttpOnly cookies.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

WebSocket open
WebSocket closed
PASS cookie is "WK-websocket-test-httponly=1; WK-websocket-test=1"
PASS successfullyParsed is true

TEST COMPLETE
